Make view part of periodical texts2

+++ src/main/java/com/theironyard/repositories/PeriodicaltextsRepository.java
@@ -3,6 +3,6 @@ package com.theironyard.repositories;
 import org.springframework.data.repository.CrudRepository;
 import com.theironyard.entities.Periodicaltexts;
 
-public interface PeriodicaltextsRepository extends CrudRepository<Periodicaltexts, Long> {
+public interface PeriodicaltextsRepository extends CrudRepository<Periodicaltexts, Integer> {
 
 }
